fix: edit page link for kb articles
This also establishes a base namespacing scheme for PHP
files to try and give some direction for future patches
and changes to websites. com\keyman\help will be the
long-term base namespace for help.keyman.com, etc. For
now we still need to require() the include files because
we need to make additional file layout changes for auto
require patterns.
