+++ _includes/includes/com.keyman.help.kb.inc.php
@@ -1,0 +1,16 @@
+<?php
+  namespace com\keyman\help\kb;
+
+  //
+  // Knowledge Base Helper Functions
+  //
+
+  function filename_from_id($id) {
+    $id = intval($id, 10);
+    return sprintf("kb%04.4d.md", $id);
+  }
+
+  function link_from_id($id) {
+    $id = intval($id, 10);
+    return "/kb/$id";
+  }
